requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0
django-dotenv==1.4.2
django-querycount==0.8.3
djangorestframework==3.16.1
drf-yasg==1.21.14
inflection==0.5.1
jwt==1.4.0
packaging==25.0
psycopg2==2.9.11
pycparser==2.23
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
